logging now takes place in logger.js
Logger class within logger.js created. it extends EventEmitter,
allowing it to emit events as a part of its log method.

the listener in app.js is now listening to an instance of the Logger
class, instead of an instance of the EventEmitter class.
this allows us to listen to the Logger class and detect when logging
takes place

changes logger.js exports to the logger class (instead of the old log
function.)